changes and fixes
